Improve state updates after light changes
The light.turn_on call will set a new state and then immediately read it
back. However, reading the state of a LIFX light right after a state
change can still return the old value.

To handle this situation we have previously delayed the update
request a little while to allow a potential state change to settle.
This is not quite right because it affects all updates, not just those
following a state change.

This commit introduces a per-light Lock to make it explicit when the
state cannot yet be trusted. This has become possible because state
changes are now run in a separate task so async_set_state no longer
has to finish quickly.

We must then do the state update ourselves. This was already done at the
end of a long transition and that code can be reused for also doing the
update at the start of a transition.