Use regular update when registering lights
This is simpler and handles initialization of all color zones in multizone
lights. We used to depend on the first periodic update to initialize all but
the first 8 zones.

Because it is now called very early, the update method can no longer
assume a valid bulb.color_zones array. Instead, we find the total length
from the response.

Also rename update_ha to update_hass because I felt like it.